Refactor bot commands into dedicated modules
