+++ src/forge_bot/commands/ping.py
@@ -1,0 +1,14 @@
+import time as time_module
+
+from telegram import Update
+from telegram.ext import ContextTypes
+
+
+async def ping(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    # Measure the round trip for editing a Telegram message as a quick bot check.
+    start = time_module.time()
+    msg = await update.message.reply_text("Calculating latency...")
+    end = time_module.time()
+
+    latencia_ms = round((end - start) * 1000)
+    await msg.edit_text(f"Latency: `{latencia_ms}ms`", parse_mode="Markdown")
